Fix of this terrible JSON bug
On May 26 at approximately 19 pm UTC, coinbase started returning 2 SAME response bodies in response to a single request. As a result, the library stopped working, since json cannot parse a double object. In fact, the API was returning an invalid json.

I made a small fix in order to cut off a copy of the answer from the original. Hopefully this will work and they won't break anything else.